agents/datetime/requirements.txt
# Core requirements
todoist-api-python>=2.0.0
python-dotenv>=0.19.0
